feat(scope-walking): YAML fromResultSet: validation + selector lowering (PR-E)
The lowered from_result_set:<id> scope kind already exists (PR-C/D); this
adds the YAML authoring surface that produces it.

- New yaml_scan.{hpp,cpp}: the four YAML line-scanners moved out of
  policy_store.cpp's anonymous namespace so a second TU can share them
  (behaviour-preserving move; policy_store re-exports them via using-
  declarations), plus a new line-anchored yaml_has_key predicate.
- New scope_yaml.{hpp,cpp}: parse_scope_block / validate_scope_block /
  lower_scope_block. Lowers fromResultSet -> from_result_set:<ref>,
  selector.platform -> ostype == "...", selector.tags -> EXISTS tag:X
  (presence; Yuzu tags are key=value), AND-composed. Enforces design §7
  rules: no assignment.managementGroups with fromResultSet, and
  assignment.mode must be static (not dynamic).
- policy_store create_policy: a scalar scope: is unchanged; a scope:
  mapping now lowers selector: (previously read as empty and silently
  stored no scope — a latent bug) and rejects fromResultSet: for policies
  (deferred to PR-E2: Policy has no owner field, the evaluator does not
  thread the result-set store, and a 1h TTL clashes with continuous eval).
- instruction_store create_definition_impl: validate the optional
  spec.scope block on every creation path (REST/import/bundled). Only
  errors when fromResultSet is present with a forbidden combo, so
  scope-less and bundled content pass untouched and boot stays safe.

Resolution stays lazy: validated at create, resolved at dispatch.
Design: docs/scope-walking-design.md §7; capability §30.
